Make the supported types an internal variable
This allows a user to override the custom types. For example, one can do

```
binarytree._ATTR_VALUE_TYPES = (float, int, np.uint8)

class CustomNode(binarytree.Node):
    pass
```

Ideally, the attributes should be members of the nodes. That would allow a user to override a lot of different options, s.a. how the values are printed, what to do with the long strings, etc.